- Hooked up alerting the match being closed - Took care of resultant circular import
