+++ src/app/sc2/domain/player.py
@@ -214,14 +214,3 @@ class PlayerDetails(object):
         player_rank = PlayerRankModel.get_or_create(player.battle_net_name, season_id)
         is_participating = season_id in player.seasons_participated
         return PlayerDetails(player, player_rank, is_participating=is_participating)
-
-
-def prettify_name(name):
-    """
-    Make the name of a user nice
-    :param name: Either a single name or a first and last name separated by a space
-    :return: The single name or the first and last name in the format "Firstname LastInitial"
-    """
-    names = name.strip().split(' ')
-    formatted_name = '{first} {last}'.format(first=names[0], last=names[1][0]) if len(names) == 2 else names[0]
-    return formatted_name
