+++ src/app/sc2/domain/slack.py
@@ -3,17 +3,16 @@ import urllib
 
 from google.appengine.api import urlfetch
 
-from app.sc2.domain.match import lookup_open_matches, get_vs_string_from_match
-from app.sc2.domain.player import prettify_name
+from app.sc2.utils.match_utils import get_vs_string_from_match
+from app.sc2.utils.player_utils import prettify_name
 
 
 # TODO: hook this up when auth stuff figured out
-def update_channel_topic_with_open_games(channel_id):
+def update_channel_topic_with_open_games(channel_id, open_matches):
     """
     Sets the channel topic to the currently open games
     Need to wait until I check out the auth stuff before this will work
     """
-    open_matches = lookup_open_matches()
     if open_matches:
         games_to_be_played = '\n'.join([get_vs_string_from_match(m) for m in open_matches])
     else:
