+++ test/sc2/domain/slack_test.py
@@ -21,14 +21,10 @@ class UpdateChannelTopicTests(TestCase):
                             new_callable=mock.PropertyMock(return_value=['Stone Cold', 'Brett Hart']))
         self.team2_patcher = mocker.start()
         self.addCleanup(mocker.stop)
-
         self.match = MatchModel()
-        mocker = mock.patch('app.sc2.domain.slack.lookup_open_matches', return_value=[self.match])
-        self.open_matches_patcher = mocker.start()
-        self.addCleanup(mocker.stop)
 
     def test_makes_call_to_slack(self):
-        slack.update_channel_topic_with_open_games(self.match)
+        slack.update_channel_topic_with_open_games('channel-id', [self.match])
         self.assertEqual(1, self.urlfetch_patcher.call_count)
