+++ test/sc2/utils/player_utils_test.py
@@ -1,4 +1,4 @@
-from app.sc2.domain import player
+from app.sc2.utils import player_utils
 
 from unittest import TestCase
 
@@ -6,12 +6,12 @@ from unittest import TestCase
 class PrettifyNameTests(TestCase):
     def test_returns_single_name_as_is(self):
         expect = 'Macho'
-        self.assertEqual(expect, player.prettify_name(expect))
+        self.assertEqual(expect, player_utils.prettify_name(expect))
 
     def test_removes_whitespace(self):
         expect = 'Macho'
-        self.assertEqual(expect, player.prettify_name('Macho '))
+        self.assertEqual(expect, player_utils.prettify_name('Macho '))
 
     def test_returns_firstname_last_initial(self):
         expect = 'Randy S'
-        self.assertEqual(expect, player.prettify_name('Randy Savage'))
+        self.assertEqual(expect, player_utils.prettify_name('Randy Savage'))
